refactored into discrete components, fixed date filtering. needs to load markers on load
